requirements.txt
gym==0.26.2
gymnasium==0.29.1
loguru==0.7.2
matplotlib==3.8.4
networkx==3.1
numpy==1.24.3
stable_baselines3==2.1.0
torch==2.3.0
tshub.egg==info
